third_party/blink/web_tests/platform/generic/http/tests/devtools/sources/debugger-ui/reveal-not-skipped-expected.txt

Tests that certain user actions in scripts panel reveal execution line.


Running: testRevealAfterPausedOnException
Showing script1 source...
Script source was shown for 'script1.js'.
Throwing exception...
Script execution paused.
Script source was shown for 'reveal-not-skipped.html'.
Reloading page...
Script execution resumed.
Page reloaded.
Showing script1 source...
Script source was shown for 'script1.js'.
Throwing exception...
Script execution paused.
Script source was shown for 'reveal-not-skipped.html'.

Running: testRevealAfterPrettyPrintWhenPaused
Throwing exception...
Showing script1 source...
Script source was shown for 'script1.js'.
Formatting...
Script source was shown for 'script1.js:formatted'.
Script execution resumed.
